backend/requirements.txt
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
rembg>=2.0.50
onnxruntime>=1.16.0
numpy>=1.24.0
aiofiles>=23.0.0
